requirements.txt
telethon
google-api-python-client
googletrans
firebase-admin
pillow
pytesseract
opencv-python